requirements_dev.txt
factory-boy==3.1.0
pre-commit==2.7.1
pytest==6.1.1
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pytest-dotenv==0.5.2
pytest-mock==3.3.1
radon==4.3.2
xenon==0.7.1
